pyproject.toml
[project]
name = "azure-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects>=1.0.0b3",
    "azure-functions>=1.24.0",
    "azure-identity>=1.15.0",
    "pydantic>=2.12.5",
    "requests>=2.32.0",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
